# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Letpai API
  version: 1.0.0
  description: API for Letpai - Bill splitting application with WhatsApp integration
  contact:
    name: Letpai Team
    email: support@letpai.com
servers:
  - url: http://localhost:3000/api/v1
  - url: /api/v1
paths:
...
  /contacts:
    get:
      summary: Get all contacts
      description: Get all contacts for current user
      operationId: getContacts
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - name: group_id
          in: query
          description: Filter by group ID
          schema:
            type: string
            format: uuid
        - name: is_favorite
          in: query
          description: Filter by favorite status
          schema:
            type: boolean
        - name: search
          in: query
          description: Search by name
          schema:
            type: string
        - name: sort_by
          in: query
          description: Sort by field
          schema:
            type: string
            enum: [name, created_at, whatsapp_number]
            default: created_at
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Contacts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create contact
      description: Create new contact
      operationId: createContact
      tags:
        - Contacts
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactRequest'
      responses:
        '201':
          description: Contact created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
...
  /contacts/import:
    post:
      summary: Import contacts
      description: Import contacts from JSON with deduplication
      operationId: importContacts
      tags:
        - Contacts
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportContactsRequest'
      responses:
        '200':
          description: Contacts imported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportContactsResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contacts/{id}:
    get:
      summary: Get contact by ID
      description: Get single contact details
      operationId: getContact
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ContactID'
      responses:
        '200':
          description: Contact retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update contact
      description: Update contact details
      operationId: updateContact
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ContactID'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactRequest'
      responses:
        '200':
          description: Contact updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'

    delete:
      summary: Delete contact
      description: Delete contact (soft delete)
      operationId: deleteContact
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ContactID'
      responses:
        '200':
          description: Contact deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /sessions/{id}/send-notifications:
    post:
      summary: Send notifications
      description: Send WhatsApp notifications to all participants
      operationId: sendNotifications
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
      responses:
        '200':
          description: Notifications sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendNotificationsResponse'
  /sessions/{id}/bulk-reminder:
    post:
      summary: Send bulk reminders
      description: Send reminder to all unpaid participants
      operationId: bulkReminder
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
      responses:
        '200':
          description: Reminders sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkReminderResponse'
...
  /participants/{participant_id}/reminder:
    post:
      summary: Send reminder
      description: Send reminder to specific participant
      operationId: sendReminder
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ParticipantID'
      responses:
        '200':
          description: Reminder sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendReminderResponse'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
                description: Seconds to wait before next reminder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "AUTH_001"
            message:
              type: string
              example: "Invalid credentials"
...
    Contact:
      type: object
      properties:
        contact_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "John Doe"
        whatsapp_number:
          type: string
          example: "+6281234567890"
        group_id:
          type: string
          format: uuid
          nullable: true
        group_name:
          type: string
          nullable: true
        group_color:
          type: string
          nullable: true
        is_favorite:
          type: boolean
          example: true
        avatar_url:
          type: string
          format: uri
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          example: "2026-03-18T10:00:00Z"
    CreateContactRequest:
      type: object
      required:
        - name
        - whatsapp_number
      properties:
        name:
          type: string
          maxLength: 100
          example: "Jane Doe"
        whatsapp_number:
          type: string
          example: "+6289876543210"
        group_id:
          type: string
          format: uuid
          nullable: true
        avatar_url:
          type: string
          format: uri
          nullable: true
        notes:
          type: string
          nullable: true
    UpdateContactRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        whatsapp_number:
          type: string
        group_id:
          type: string
          format: uuid
          nullable: true
        is_favorite:
          type: boolean
        avatar_url:
          type: string
          format: uri
          nullable: true
        notes:
          type: string
          nullable: true
    ImportContactsRequest:
      type: object
      required:
        - contacts
      properties:
        contacts:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ImportContact'
        group_id:
          type: string
          format: uuid
          nullable: true
    ImportContact:
      type: object
      required:
        - name
        - whatsapp_number
      properties:
        name:
          type: string
          maxLength: 100
        whatsapp_number:
          type: string
    ImportContactsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        imported_count:
          type: integer
          example: 5
        skipped_count:
          type: integer
          example: 2
        skipped_duplicates:
          type: array
          items:
            type: string
          example:
            - "+6281234567890"
            - "+6289876543210"
        message:
          type: string
    ContactResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Contact'
    ContactsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 50
...
    SendNotificationsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "3 notifications sent"
        data:
          type: array
          items:
            type: object
            properties:
              participant_id:
                type: string
                format: uuid
              whatsapp_message_id:
                type: string
                example: "msg_123"
              sent_at:
                type: string
                format: date-time
                example: "2026-03-18T14:00:00Z"
    SendReminderResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Reminder sent to John Doe"
        data:
          type: object
          properties:
            whatsapp_message_id:
              type: string
              example: "msg_456"
            next_available_at:
              type: string
              format: date-time
              example: "2026-03-19T14:00:00Z"
    BulkReminderResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "2 reminders sent"
        skipped:
          type: array
          items:
            type: object
            properties:
              participant_id:
                type: string
                format: uuid
              reason:
                type: string
                example: "Rate limit: wait 22h"
    RateLimitError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Rate limit exceeded. Wait 22h before next reminder."
        retry_after:
          type: integer
          example: 79200
          description: Seconds to wait
...
  parameters:
    ContactID:
      name: id
      in: path
      required: true
      description: Contact ID
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...
    SessionID:
      name: id
      in: path
      required: true
      description: Session ID
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    ParticipantID:
      name: participant_id
      in: path
      required: true
      description: Participant ID
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...